Move getWalletLink function into the routing helper
